renamed directory ttcr to cutils to be more general
